Fork code for reproducing ICN'13 results from Icarus simulator
This commit creates a new subfolder that will contain all the code
for reproducing the results of the ICN'13 hash-routing paper.

It would be more logical to create a branch instead of a subfolder.
However, I decided to do so to because it would be easier for users
to just download the master branch and have everything there.